Split TripleAForumPoster into 2 classes for future migration
